requirements.txt
langgraph
langchain
langchain-groq
streamlit
requests
beautifulsoup4
newspaper3k
python-dotenv
tavily-python
pydantic
typing-extensions
matplotlib
plotly
wordcloud
textstat
dotenv
